refactor: integration 패키지의 이름을 email로 변경
