+++ src/main/java/com/gdschongik/gdsc/domain/email/dao/VerificationCodeAndEmailRepository.java
@@ -1,6 +1,6 @@
-package com.gdschongik.gdsc.domain.integration.dao;
+package com.gdschongik.gdsc.domain.email.dao;
 
-import com.gdschongik.gdsc.domain.integration.domain.VerificationCodeAndEmail;
+import com.gdschongik.gdsc.domain.email.domain.VerificationCodeAndEmail;
 import org.springframework.data.repository.CrudRepository;
 
 public interface VerificationCodeAndEmailRepository extends CrudRepository<VerificationCodeAndEmail, String> {}
