+++ src/main/java/com/gdschongik/gdsc/domain/email/util/VerificationLinkUtil.java
@@ -1,13 +1,12 @@
-package com.gdschongik.gdsc.domain.integration.util;
+package com.gdschongik.gdsc.domain.email.util;
 
-import static com.gdschongik.gdsc.domain.integration.constant.UnivMailVerificationConstant.VERIFY_EMAIL_API_ENDPOINT;
-import static com.gdschongik.gdsc.domain.integration.constant.UnivMailVerificationConstant.VERIFY_EMAIL_QUERY_STRING_KEY;
 import static com.gdschongik.gdsc.global.common.constant.EnvironmentConstant.DEV;
 import static com.gdschongik.gdsc.global.common.constant.EnvironmentConstant.PROD;
 import static com.gdschongik.gdsc.global.common.constant.UrlConstant.DEV_CLIENT_URL;
 import static com.gdschongik.gdsc.global.common.constant.UrlConstant.LOCAL_VITE_CLIENT_SECURE_URL;
 import static com.gdschongik.gdsc.global.common.constant.UrlConstant.PROD_CLIENT_URL;
 
+import com.gdschongik.gdsc.domain.email.constant.UnivMailVerificationConstant;
 import com.gdschongik.gdsc.global.util.EnvironmentUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -19,7 +18,8 @@ public class VerificationLinkUtil {
     private final EnvironmentUtil environmentUtil;
 
     public String createLink(String verificationCode) {
-        String verifyEmailApiEndpoint = String.format(VERIFY_EMAIL_API_ENDPOINT, VERIFY_EMAIL_QUERY_STRING_KEY);
+        String verifyEmailApiEndpoint = String.format(
+            UnivMailVerificationConstant.VERIFY_EMAIL_API_ENDPOINT, UnivMailVerificationConstant.VERIFY_EMAIL_QUERY_STRING_KEY);
         return getClientUrl() + verifyEmailApiEndpoint + verificationCode;
     }
