+++ src/main/java/com/gdschongik/gdsc/domain/email/util/VerificationMailContentWriter.java
@@ -1,4 +1,4 @@
-package com.gdschongik.gdsc.domain.integration.util;
+package com.gdschongik.gdsc.domain.email.util;
 
 import java.time.Duration;
 import org.springframework.stereotype.Component;
