+++ src/main/java/com/gdschongik/gdsc/global/util/email/JavaMailSender.java
@@ -1,4 +1,4 @@
-package com.gdschongik.gdsc.global.util.mail;
+package com.gdschongik.gdsc.global.util.email;
 
 import static com.gdschongik.gdsc.global.common.constant.EmailConstant.SENDER_ADDRESS;
 import static com.gdschongik.gdsc.global.common.constant.EmailConstant.SENDER_PERSONAL;
