+++ src/test/java/com/gdschongik/gdsc/domain/email/HongikUnivEmailValidatorTest.java
@@ -1,11 +1,11 @@
-package com.gdschongik.gdsc.domain.integration;
+package com.gdschongik.gdsc.domain.email;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.gdschongik.gdsc.global.exception.CustomException;
 import com.gdschongik.gdsc.global.exception.ErrorCode;
-import com.gdschongik.gdsc.global.util.mail.HongikUnivEmailValidator;
+import com.gdschongik.gdsc.global.util.email.HongikUnivEmailValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
